Add grouping information of the WDI data to a metric summary

Add grouping information of the WDI data to a metric summary

Usage

add_group_info(metric_summary, wdi_data)

Arguments

metric_summary

A data frame containing the calculated diagnostic indices generated by any of the following functions: compute_variation, compute_trend_shape_features, compute_temporal_features, or compute_diagnostic_indices

wdi_data

A data frame of the indicator data generated by get_wdi_data

Value

A data frame containing the calculated diagnostic indices and the grouping variables in the WDI data set.

Examples

pm_diagnostic_metrics <- compute_diagnostic_indices(pm_data, group_var = "region")
#> Registered S3 method overwritten by 'tsibble':
#>   method               from 
#>   as_tibble.grouped_df dplyr
pm_diagnostic_metrics_group <- add_group_info(metric_summary = pm_diagnostic_metrics,pm_data)