Compute sequential temporal features

Calculates number of crossing points, longest flat spot using the feasts package functionality and an additional time series feature - autocorrelation.

Usage

compute_temporal_features(wdi_data, index = NULL)

Arguments

wdi_data

A data frame of the indicator data generated by get_wdi_data

index

An optional character string specifying the indicator code Defaults to NULL

Value

A data frame with columns country, crossing_points, flat_spot, and acf.

Examples

pm_temporal <- compute_temporal_features(pm_data)