Extract valid data from the WDI data Reports countries with no data points as well as years for which no data are available.

Extract valid data from the WDI data Reports countries with no data points as well as years for which no data are available.

Usage

get_valid_data(wdi_data, index = NULL)

Arguments

wdi_data

A data frame of the indicator data generated by get_wdi_data

index

An optional character string specifying the indicator code Defaults to NULL

Value

A tibble with the valid data for the provided WDI indicator data set and a detailed report of missing entries.

Examples

get_valid_data(pm_data)
#> The 17 countries listed below had no available data and were excluded:
#>  Aruba
#> - British Virgin Islands
#> - Cayman Islands
#> - Channel Islands
#> - Curacao
#> - Faroe Islands
#> - French Polynesia
#> - Gibraltar
#> - Hong Kong SAR, China
#> - Isle of Man
#> - Kosovo
#> - Liechtenstein
#> - Macao SAR, China
#> - New Caledonia
#> - Sint Maarten (Dutch part)
#> - St. Martin (French part)
#> - Turks and Caicos Islands
#> 
#> The 34 year(s) listed below had no available data and were excluded:
#>  1960, 	1961, 	1962, 	1963, 	1964, 	1965, 	1966, 	1967, 	1968, 	1969, 	1970, 	1971, 	1972, 	1973, 	1974, 	1975, 	1976, 	1977, 	1978, 	1979, 	1980, 	1981, 	1982, 	1983, 	1984, 	1985, 	1986, 	1987, 	1988, 	1989, 	2021, 	2022, 	2023, 	2024 
#> # A tibble: 6,169 × 13
#>    country iso2c iso3c  year EN.ATM.PM25.MC.M3 status lastupdated region capital
#>    <chr>   <chr> <chr> <int>             <dbl> <chr>  <chr>       <chr>  <chr>  
#>  1 Afghan… AF    AFG    1990             64.2  ""     2025-07-01  South… Kabul  
#>  2 Albania AL    ALB    1990             23.0  ""     2025-07-01  Europ… Tirane 
#>  3 Algeria DZ    DZA    1990             22.4  ""     2025-07-01  Middl… Algiers
#>  4 Americ… AS    ASM    1990              6.43 ""     2025-07-01  East … Pago P…
#>  5 Andorra AD    AND    1990             16.8  ""     2025-07-01  Europ… Andorr…
#>  6 Angola  AO    AGO    1990             23.8  ""     2025-07-01  Sub-S… Luanda 
#>  7 Antigu… AG    ATG    1990             19.3  ""     2025-07-01  Latin… Saint …
#>  8 Argent… AR    ARG    1990             19.0  ""     2025-07-01  Latin… Buenos…
#>  9 Armenia AM    ARM    1990             28.2  ""     2025-07-01  Europ… Yerevan
#> 10 Austra… AU    AUS    1990              7.76 ""     2025-07-01  East … Canber…
#> # ℹ 6,159 more rows
#> # ℹ 4 more variables: longitude <chr>, latitude <chr>, income <chr>,
#> #   lending <chr>