PM2.5 air pollution data

This data set contains the mean annual exposure levels to ambient PM2.5 air pollution across various countries, measured in micrograms per cubic meter.

Usage

pm_data

Format

A data frame with 13,910 observations with 13 variables

country

Country name (character)

iso2c

2-letter ISO country code (character)

iso3c

3-letter ISO country code (character)

year

Calendar year representing the time index of the observation (integer)

EN.ATM.PM25.MC.M3

Observational values for the specified indicator code (numeric)

status

An empty variable meant to indicate the operational status of variables (character)

lastupdated

Timestamp that indicates the most recent update of the indicator date (character)

region

Geographical region variable (character)

capital

Name of the capital city of each country (character)

longitude

Geographic coordinate that measures the longitude of the city (character)

latitude

Geographic coordinate that measures the latitude of the city (character)

income

World Bank income classification variable (character)

lending

World Bank income classification variable (character)

Source

World Development Indicator, using the WDI R package

Examples

data(pm_data)

head(pm_data)
#>       country iso2c iso3c year EN.ATM.PM25.MC.M3 status lastupdated     region
#> 1 Afghanistan    AF   AFG 2013          73.13182         2025-07-01 South Asia
#> 2 Afghanistan    AF   AFG 2012          70.92232         2025-07-01 South Asia
#> 3 Afghanistan    AF   AFG 2014          77.14373         2025-07-01 South Asia
#> 4 Afghanistan    AF   AFG 2009          51.00144         2025-07-01 South Asia
#> 5 Afghanistan    AF   AFG 2008          54.19114         2025-07-01 South Asia
#> 6 Afghanistan    AF   AFG 2007          58.08379         2025-07-01 South Asia
#>   capital longitude latitude     income lending
#> 1   Kabul   69.1761  34.5228 Low income     IDA
#> 2   Kabul   69.1761  34.5228 Low income     IDA
#> 3   Kabul   69.1761  34.5228 Low income     IDA
#> 4   Kabul   69.1761  34.5228 Low income     IDA
#> 5   Kabul   69.1761  34.5228 Low income     IDA
#> 6   Kabul   69.1761  34.5228 Low income     IDA